# Speed up Semantic.Git.catFile and avoid crashes on non-UTF-8 input.