--- a/bench/Main.hs
+++ b/bench/Main.hs
@@ -1,0 +1,9 @@
+module Main where
+
+import Gauge
+import qualified Evaluation
+
+main :: IO ()
+main = defaultMain
+  [ Evaluation.benchmarks
+  ]