--- a/semantic-source/CHANGELOG.md
+++ b/semantic-source/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.0.2
+
+- Adds an `NFData` instance for `Source`.
+
 # 0.0.0.1
 
 - Loosens the upper bound on `hashable`.