--- a/semantic-source/CHANGELOG.md
+++ b/semantic-source/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.0.2
+# 0.0.1.0
 
 - Adds an `NFData` instance for `Source`.