--- a/semantic.cabal
+++ b/semantic.cabal
@@ -68,6 +68,7 @@ common dependencies
                      , semilattices ^>= 0.0.0.3
                      , shelly >= 1.5 && <2
                      , streaming ^>= 0.2.2.0
+                     , streaming-bytestring ^>= 0.1.6
                      , text ^>= 1.2.3.1
                      , these >= 0.7 && <1
                      , unix ^>= 2.7.2.2
@@ -311,6 +312,7 @@ library
                      , semantic-tags ^>= 0
                      , semigroupoids ^>= 5.3.2
                      , split ^>= 0.2.3.3
+                     , streaming-process ^>= 0.1
                      , stm-chans ^>= 3.0.0.4
                      , template-haskell ^>= 2.14
                      , time ^>= 1.8.0.2
@@ -410,15 +412,16 @@ test-suite parse-examples
                      , foldl ^>= 1.4.5
                      , resourcet ^>= 1.2
                      , streaming
-                     , streaming-bytestring ^>= 0.1.6
+                     , streaming-bytestring
                      , tasty
                      , tasty-hunit
 
 benchmark evaluation
   import:              haskell, dependencies, executable-flags
-  hs-source-dirs:      bench/evaluation
+  hs-source-dirs:      bench
   type:                exitcode-stdio-1.0
   main-is:             Main.hs
+  other-modules:       Evaluation
   ghc-options:         -static
   build-depends:       base
                      , algebraic-graphs