When using `semantic parse --gitDir` on repos containing non-UTF-8 source files (like [this one](https://github.com/drupal/drupal/blob/8.8.x/core/tests/Drupal/Tests/Core/Asset/js_test_files/latin_9.js)), we were observing `semantic` terminating unexpectedly with this message:

```
*** Exception: 
Ran commands: 
git -C /tmp/haridusportaal/.git cat-file -p a488429fe5c9ada71e2b8dcaaac821fe488c0d44
which git

Exception: fd:15: hGetLine: invalid argument (invalid byte sequence)
```

We cannot handle this case perfectly, because `tree-sitter` doesn’t work with anything but UTF-8/16. However, we can do much better than just crashing the whole program. `shelly` is attempting to convert the output of this git process to UTF-8 immediately, before sending it to tree-sitter; if we avoid `shelly`, we can stream the output directly into a `Source` without crashing. We will still fail to return results for the files in question, since `tree-sitter` can’t make sense of them, but those failures will not cause the whole invocation to terminate.

In addition, we can get more performance out of this hot path by avoiding the overhead of `shelly` converting the `git` output to `Text` and then back to a `ByteString` with `Source.fromText`. We can just stream the output directly into a strict `ByteString` and avoid a pointless roundtrip.

I benchmarked this implementation (using `streaming-bytestring`) against implementation `shelly` and `process-extras`, factoring in the fact that `shelly` has to call `Source.fromText` to convert to `ByteString`. The results show that the `streaming` method is the fastest:

```
benchmarked utility/git/catFile/shelly
time                 8.523 ms   (7.920 ms .. 9.004 ms)
                     0.945 R²   (0.859 R² .. 0.991 R²)
mean                 9.019 ms   (8.724 ms .. 9.917 ms)
std dev              1.248 ms   (546.1 μs .. 2.469 ms)
variance introduced by outliers: 68% (severely inflated)

benchmarked utility/git/catFile/process-extras
time                 8.300 ms   (8.073 ms .. 8.497 ms)
                     0.993 R²   (0.983 R² .. 0.998 R²)
mean                 8.301 ms   (8.192 ms .. 8.502 ms)
std dev              429.2 μs   (270.7 μs .. 748.8 μs)
variance introduced by outliers: 25% (moderately inflated)

benchmarked utility/git/catFile/streaming-bytestring
time                 7.692 ms   (7.191 ms .. 8.286 ms)
                     0.981 R²   (0.969 R² .. 0.998 R²)
mean                 7.774 ms   (7.659 ms .. 7.936 ms)
std dev              375.9 μs   (289.9 μs .. 579.5 μs)
variance introduced by outliers: 21% (moderately inflated)
```